cleanup code and extract sync logic into hook
