+++ react-main-container/.eslintrc.js
