+++ react-main-container/package.json
@@ -5,6 +5,10 @@
     "build": "webpack --mode production",
     "build:dev": "webpack --mode development",
     "build:start": "cd dist && PORT=4000 npx serve",
+    "lint": "eslint",
+    "lint-i18n": "eslint --ext .json  --format node_modules/eslint-plugin-i18n-json/formatter.js public/locales/",
+    "lint-i18n:fix": "eslint --fix --ext .json  --format node_modules/eslint-plugin-i18n-json/formatter.js public/locales/",
+    "lint:fix": "eslint --fix",
     "start": "set NODE_OPTIONS=--openssl-legacy-provider && webpack serve --open --mode development",
     "start:live": "webpack serve --open --mode development --live-reload --hot"
   },
@@ -17,8 +21,14 @@
     "@babel/preset-typescript": "^7.10.4",
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
     "babel-loader": "^8.2.2",
     "css-loader": "^6.3.0",
+    "eslint": "^8.37.0",
+    "eslint-plugin-i18n-json": "^4.0.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.3.2",
     "style-loader": "^3.3.0",
     "typescript": "^5.0.2",
