+++ react-main-container/public/locales/en/common.json
@@ -1,5 +1,5 @@
 {
-  "NAME": "Name",
   "FRAMEWORK": "Framework",
-  "LANGUAGE": "Language"
+  "LANGUAGE": "Language",
+  "NAME": "Name"
 }
