+++ react-main-container/public/locales/fr/app.json
@@ -1,3 +1,3 @@
 {
-  "TITLE": "Bienvenue dans l'application principale !"
+  "TITLE": "Bienvenue dans l'application principale!"
 }
