+++ react-main-container/public/locales/fr/common.json
@@ -1,5 +1,5 @@
 {
-  "NAME": "Nom",
   "FRAMEWORK": "Cadre",
-  "LANGUAGE": "Langue"
+  "LANGUAGE": "Langue",
+  "NAME": "Nom"
 }
