+++ react-main-container/public/locales/i18n.ts
@@ -1,5 +1,4 @@
-// import { initReactI18next } from 'react-i18next';
-import i18n, { createInstance } from 'i18next';
+import { createInstance } from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import HttpApi from 'i18next-http-backend';
 
@@ -11,8 +10,6 @@ appI18n
   // detect user language
   // learn more: https://github.com/i18next/i18next-browser-languageDetector
   .use(LanguageDetector)
-  // // pass the i18n instance to react-i18next.
-  // .use(initReactI18next)
   // init i18next
   // for all options read: https://www.i18next.com/overview/configuration-options
   .init({
@@ -30,7 +27,6 @@ appI18n
     },
     react: {
       transSupportBasicHtmlNodes: true,
-      // transKeepBasicHtmlNodesFor: ['br', 'strong', 'i'],
     },
   });
