+++ react-mfe/package.json
@@ -29,7 +29,10 @@
     "@typescript-eslint/parser": "^5.57.1",
     "babel-loader": "^8.2.2",
     "css-loader": "^6.3.0",
+    "eslint": "^8.37.0",
     "eslint-plugin-i18n-json": "^4.0.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.3.2",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
@@ -40,9 +43,6 @@
     "webpack-dev-server": "^4.3.1"
   },
   "dependencies": {
-    "eslint": "^8.37.0",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
     "i18next": "^22.4.13",
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-http-backend": "^2.2.0",
