+++ react-mfe/public/locales/en/host-common.json
@@ -1,0 +1,1 @@
+{}
