+++ react-mfe/public/locales/fr/host-common.json
@@ -1,0 +1,1 @@
+{}
