+++ react-mfe/src/@types/i18next.d.ts
@@ -1,8 +1,5 @@
 import '../../public/locales/i18n'; // import in regular types
 
-// name must be i18next.d.ts - https://github.com/i18next/react-i18next/tree/master/example/react-typescript/simple-multi-namespaces
-// documentation suggesting using react-i18next.d.ts does not work
-
 declare module 'i18next' {
   interface CustomTypeOptions {
     defaultNS: 'mfe';
