+++ react-mfe/src/tests/Microfrontend.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import '@testing-library/jest-dom';
-import { renderWithTestI18n, screen } from './test-utils';
-import enMfeTranslations from '../../public/locales/en/mfe.json';
 
-import { Information } from '../Microfrontend';
+import { renderWithTestI18n, screen } from './test-utils';
+import { Information } from '../components/Information';
+import enMfeTranslations from '../../public/locales/en/mfe.json';
 
 describe('Microfrontend Information', () => {
   beforeEach(() => {
@@ -12,6 +12,7 @@ describe('Microfrontend Information', () => {
 
   it('shows the introductory text', () => {
     screen.getByText(enMfeTranslations.GREETING);
+    screen.getByText('The MFE is able to use translations from the host such as: NAME'); // for translations from the host, we can only assert on the key
     screen.getByText(
       "What about interpolation? If it's simple enough you can directly use the `t` function and pass in variables. If you need to embed HTML, then you will need to use the `Trans` component. Using `transSupportBasicHtmlNodes` we can simplify some boilerplate around HTML tags typically used to style a subset of the text like bold and italics. For more complicated elements, we need to pass in a `components` prop. See following examples:"
     );
@@ -24,27 +25,11 @@ describe('Microfrontend Information', () => {
 });
 
 /*
+// for a set up where you are mocking i18n and translations
 describe('Microfrontend Information - Mock I18n', () => {
   // enable `react-i18next.js` in mocks folder for these tests to run
   beforeEach(() => {
     render(<Information />);
-
-    // jest.mock('react-i18next', () => ({
-    //   // this mock makes sure any components using the translate hook can use it without a warning being shown
-    //   useTranslation: () => {
-    //     return {
-    //       t: (str: string) => str,
-    //       i18n: {
-    //         changeLanguage: () => new Promise(() => ({})),
-    //       },
-    //     };
-    //   },
-    //   initReactI18next: {
-    //     type: '3rdParty',
-    //     init: () => ({}),
-    //   },
-    // }));
-    // jest.resetAllMocks();
   });
 
   it('shows the introductory text', () => {
